* master6: (27 commits)
  docs(knowledge): remove build commands and single-repo compile notes
  update: 更新文件 bundle.json
  delete unused header
  add codeonwer for depth component
  fix: tian'jia添加数据判空
  update: 更新文件 hdilayer_test.cpp
  tdd fix
  fix format
  delete marshalling for SpatialEffectPara
  差异消除 Signed-off-by: m00472246 <majingtao1@huawei.com>
  update
  fix code check
  fix code check
  fix warning
  update: 更新文件 rs_profiler_hrp_service.cpp
  update: 更新文件 rs_profiler_hrp_service.cpp
  update: 更新文件 rs_profiler.cpp add 2026063004580
  fix format
  depth component
  Add version isolation for letterSpacing and wordSpacing feature
  ...

+++ rosen/modules/render_service/main/render_server/transaction/rs_client_to_service_connection.cpp
@@ -2614,7 +2614,7 @@ RetCodeHrpService RSClientToServiceConnection::ProfilerServiceOpenFile(const Hrp
         return RET_HRP_SERVICE_ERR_INVALID_PARAM;
     }
 
-    return RSProfiler::HrpServiceOpenFile(dirInfo, fileName, flags, outFd);
+    return RSProfiler::HrpServiceOpenFile(dirInfo, fileName, static_cast<uint32_t>(flags), outFd);
 #else
     outFd = -1;
     return RET_HRP_SERVICE_ERR_UNSUPPORTED;
